# public/docs/openapi.yaml
openapi: 3.0.0
info:
  title: 'Cairo Store API'
  description: "كل نقاط المتجر ولوحة التحكم.\r\n\r\nملاحظة عامة على الأخطاء: نقاط JSON في هذا المشروع تُرجع 200 حتى عند\r\nفشل التحقق، والنتيجة تُقرأ من الحقل success في الجسم لا من كود HTTP.\r\nالاستثناءات الوحيدة هي 302 (تحويل لتسجيل الدخول) و403 (صلاحية ناقصة)\r\nو404 (راوت غير موجود). هذا سلوك قائم يعتمد عليه الـfrontend، ووُثّق\r\nهنا كما هو لا كما ينبغي أن يكون."
  version: 1.1.0
servers:
  -
    url: 'http://localhost/STORE/public'
    description: 'Local / Development Server'
paths:
...
  /products:
    get:
      tags:
        - 'Store - Products'
      summary: 'Visible product list with pagination'
      description: 'Search, sorting and price filtering all happen in the browser (js/features/products-catalog.js), so there are no filter parameters here.'
      operationId: 3766bdcbe4e6257669b4818e48e63431
      parameters:
        -
          name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          $ref: '#/components/responses/HtmlPage'
        '404':
          $ref: '#/components/responses/NotFoundPage'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /product:
    get:
      tags:
        - 'Store - Products'
      summary: 'Product details page'
      description: "Redirects home when the id is missing or the product does not exist. The displayed variant is chosen from the visitor's gender (pickDisplayVariant)."
      operationId: 71949b3453b48a350070f790deaba7d1
      parameters:
        -
          name: id
          in: query
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: 'HTML page'
        '302':
          description: 'Redirect home — missing id or unknown product'
    post:
      tags:
        - 'Store - Products'
      summary: "Save the user's product review, then render the page"
      description: 'The same method serves GET and POST. Requires a logged-in user and a CSRF token. An admin in store-browsing mode cannot leave a review.'
      operationId: 26deaeebf822fcad11cd8cdfce47d5e8
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              required:
                - submit_review
                - rating
                - csrf_token
              properties:
                submit_review:
                  type: string
                rating:
                  type: integer
                  maximum: 5
                  minimum: 1
                comment:
                  type: string
                csrf_token:
                  type: string
              type: object
      responses:
        '200':
          $ref: '#/components/responses/HtmlPage'
        '404':
          $ref: '#/components/responses/NotFoundPage'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
      security:
        -
          userSessionAuth: []
...
components:
...
  responses:
...
    HtmlPage:
      description: 'صفحة HTML كاملة.'
      content:
        text/html: {  }
...
    NotFoundPage:
      description: 'راوت غير مسجَّل أو مورد غير موجود. صفحة HTML من ErrorPage::notFound.'
      content:
        text/html: {  }
    ServiceUnavailable:
      description: "عطل تقني يمنع إكمال الطلب — أبرزه فشل الاتصال بقاعدة البيانات.\r\n\r\n503 لا 500 عمداً: الخدمة غير متاحة مؤقتاً لا «خطأ في الخادم»،\r\nوالفرق يهمّ محرّكات البحث وأدوات المراقبة. التفاصيل تذهب إلى سجلّ\r\nالأخطاء ولا تُطبع للزائر أبداً — رسالة PDO تحمل اسم المضيف واسم\r\nالقاعدة واسم المستخدم."
      content:
        text/html: {  }
...
  securitySchemes:
    adminSessionAuth:
      type: apiKey
      description: 'PHP session cookie — يُنشأ تلقائياً عند تسجيل الدخول ويُرسل مع كل طلب محمي'
      name: admin_session
      in: cookie
    userSessionAuth:
      type: apiKey
      description: "جلسة المستخدم العادي — منفصلة تماماً عن admin_session بالاسم\r\nوالمحتوى، فلا يمكن الوصول لبيانات الأدمن من جلسة مستخدم ولا العكس\r\n(راجع isUser/isAdmin في auth_helper.php)."
      name: PHPSESSID
      in: cookie
...